feat: create an aggregation to resume cards
